+++ src/snowflake/cli/api/commands/flags.py
@@ -469,13 +469,17 @@ def experimental_option(
 
 def identifier_argument(sf_object: str, example: str) -> typer.Argument:
     return typer.Argument(
-        ..., help=f"Identifier of the {sf_object}. For example: {example}"
+        ...,
+        help=f"Identifier of the {sf_object}. For example: {example}",
+        show_default=False,
     )
 
 
 def execution_identifier_argument(sf_object: str, example: str) -> typer.Argument:
     return typer.Argument(
-        ..., help=f"Execution identifier of the {sf_object}. For example: {example}"
+        ...,
+        help=f"Execution identifier of the {sf_object}. For example: {example}",
+        show_default=False,
     )