+++ src/snowflake/cli/plugins/object/command_aliases.py
@@ -1,0 +1,51 @@
+from __future__ import annotations
+
+from typing import Tuple
+
+import typer
+from snowflake.cli.api.commands.snow_typer import SnowTyper
+from snowflake.cli.api.constants import ObjectType
+from snowflake.cli.plugins.object.commands import (
+    describe,
+    drop,
+    list_,
+    scope_option,  # noqa: F401
+)
+
+
+def add_object_command_aliases(
+    app: SnowTyper,
+    object_type: ObjectType,
+    name_argument: typer.Argument,
+    like_option: typer.Option,
+    scope_option: typer.Option,
+):
+    @app.command("list", requires_connection=True)
+    def list_cmd(
+        like: str = like_option,
+        scope: Tuple[str, str] = scope_option,
+        **options,
+    ):
+        list_(object_type=object_type.value.cli_name, like=like, scope=scope, **options)
+
+    list_cmd.__doc__ = f"Lists all available {object_type.value.sf_plural_name}."
+
+    @app.command("drop", requires_connection=True)
+    def drop_cmd(name: str = name_argument, **options):
+        drop(
+            object_type=object_type.value.cli_name,
+            object_name=name,
+            **options,
+        )
+
+    drop_cmd.__doc__ = f"Drops {object_type.value.sf_name} with given name."
+
+    @app.command("describe", requires_connection=True)
+    def describe_cmd(name: str = name_argument, **options):
+        describe(
+            object_type=object_type.value.cli_name,
+            object_name=name,
+            **options,
+        )
+
+    describe_cmd.__doc__ = f"Provides description of {object_type.value.sf_name}."